% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osg_clean.r
\name{osg_clean}
\alias{osg_clean}
\title{This function cleans directories of given files and sub-directories on the Open Science Grid (OSG).}
\usage{
osg_clean(
  session = NULL,
  unix_name,
  login_node,
  rsa_keyfile = NULL,
  rsa_passphrase = NULL,
  remote_dirs,
  clean_files = NULL,
  clean_subdirs = NULL,
  verbose = TRUE
)
}
\arguments{
\item{session}{ssh connection created by \link[ssgrid]{osg_connect}.}

\item{unix_name}{Character string giving OSG unix login name.}

\item{login_node}{Character string giving OSG login node (e.g., login05.osgconnect.net).}

\item{rsa_keyfile}{Path to private key file. Must be in OpenSSH format (see details). Default is NULL. See \link[ssh]{ssh_connect} for more details.}

\item{rsa_passphrase}{Either a string or a callback function for password prompt. Default is NULL. See \link[ssh]{ssh_connect} for more details.}

\item{remote_dirs}{Character vector giving path of directories to clean on OSG login node.}

\item{clean_files}{Character vector giving files to remove from remote_dirs. Defaults to all files except for End.tar.gz.}

\item{clean_subdirs}{Character vector giving path of directories to delete on OSG login node.}

\item{verbose}{Boolean denoting if function details should be printed.}
}
\value{
Returns 0 on exit.
}
\description{
This function cleans directories of given files and sub-directories on the Open Science Grid (OSG).
}
