% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osg_condor_submit_create.r
\name{osg_condor_submit_create}
\alias{osg_condor_submit_create}
\title{This function writes a condor_submit script to be executed on the open science grid.}
\usage{
osg_condor_submit_create(
  session = NULL,
  unix_name,
  login_node,
  rsa_keyfile = NULL,
  rsa_passphrase = NULL,
  local_shell_path = NULL,
  remote_shell_path = "scripts/condor_submit/",
  file_name = "condor.sub",
  c_executable = "scripts/bash/wrapper.sh",
  c_input_files = c("Start.tar.gz"),
  c_output_files = c("End.tar.gz"),
  c_project = NULL,
  c_memory = "2GB",
  c_disk = "2GB",
  c_target_dir_path = NULL,
  c_singularity = NULL,
  overwrite = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{rsa_passphrase}{}
}
\description{
This function saves a copy of locally, uploads a copy to osg, and modifies permissions so that it can be executed remotely
}
