% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osg_connect.r
\name{osg_connect}
\alias{osg_connect}
\title{This function opens an ssh connection to the OSG.}
\usage{
osg_connect(unix_name, login_node, rsa_keyfile = NULL, rsa_passphrase = NULL)
}
\arguments{
\item{rsa_passphrase}{}
}
\value{
Returns an object created by ssh::ssh_connect.
}
\description{
If you have a passphrase associated with your rsa private key please enter it in the 'password' section of the dialogue box that opens.
This is function is a wrapper for ssh::ssh_connect to connect to the OSG login node via ssh
}
\details{
Please use ?ssh::ssh_connect for more information.
}
