% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osg_monitor.r
\name{osg_monitor}
\alias{osg_monitor}
\title{This function monitors an Open Science Grid (OSG) job submission}
\usage{
osg_monitor(
  session = NULL,
  unix_name,
  login_node,
  rsa_keyfile = NULL,
  rsa_passphrase = NULL
)
}
\arguments{
\item{session}{ssh connection created by \link{osg_connect}.}

\item{unix_name}{Character string giving OSG unix login name.}

\item{login_node}{Character string giving OSG login node (e.g., login05.osgconnect.net).}

\item{rsa_keyfile}{Path to private key file. Must be in OpenSSH format (see details). Default is NULL. See \link[ssh]{ssh_connect} for more details.}

\item{rsa_passphrase}{Either a string or a callback function for password prompt. Default is NULL. See \link[ssh]{ssh_connect} for more details.}
}
\value{
Returns the output from executing the \href{https://htcondor.readthedocs.io/en/latest/man-pages/condor_q.html}{condor_q} command on the OSG.
}
\description{
It is a wrapper around the \href{https://htcondor.readthedocs.io/en/latest/man-pages/condor_q.html}{condor_q} function
}
