% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osg_multi_copy.r
\name{osg_multi_copy}
\alias{osg_multi_copy}
\title{This function copies files to all specified directories on the OSG}
\usage{
osg_multi_copy(
  session = NULL,
  unix_name,
  login_node,
  rsa_keyfile = NULL,
  rsa_passphrase = NULL,
  remote_source_path,
  files_to_copy,
  remote_paste_path,
  verbose = TRUE
)
}
\arguments{
\item{rsa_passphrase}{}
}
\description{
Please see ?help for osg_connect and ssh::ssh_connect for more information.
}
