% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osg_r_create.r
\name{osg_r_create}
\alias{osg_r_create}
\title{This function writes an R script which can be executed within the wrapper.sh to set-up and run diagnostics using existing r4ss code.}
\usage{
osg_r_create(
  session = NULL,
  unix_name,
  login_node,
  rsa_keyfile = NULL,
  rsa_passphrase = NULL,
  local_shell_path = NULL,
  remote_shell_path = "scripts/r/",
  diagnostic_type = NULL,
  overwrite = TRUE,
  verbose = TRUE,
  retro_n_years = NULL,
  r0_maxdiff = NULL,
  r0_step = NULL
)
}
\arguments{
\item{rsa_passphrase}{}
}
\description{
This function saves a copy locally, uploads a copy to osg, and modifies permissions so that it can be executed remotely
}
