% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_control-data.r
\docType{data}
\name{ss_control}
\alias{ss_control}
\title{Example Stock Synthesis control file. Simulated data loosely based on the Pacific Hake stock assessment.}
\format{
Named list created using \link[r4ss]{SS_readctl}.
}
\usage{
ss_control
}
\description{
Example Stock Synthesis control file. Simulated data loosely based on the Pacific Hake stock assessment.
}
\keyword{datasets}
