% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_data-data.r
\docType{data}
\name{ss_data}
\alias{ss_data}
\title{Example Stock Synthesis data file. Simulated data loosely based on the Pacific Hake stock assessment.}
\format{
Named list created using \link[r4ss]{SS_readdat}.
}
\usage{
ss_data
}
\description{
Example Stock Synthesis data file. Simulated data loosely based on the Pacific Hake stock assessment.
}
\keyword{datasets}
