% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_forecast-data.r
\docType{data}
\name{ss_forecast}
\alias{ss_forecast}
\title{Example Stock Synthesis forecast file. Simulated data loosely based on the Pacific Hake stock assessment.}
\format{
Named list created using \link[r4ss]{SS_readforecast}.
}
\usage{
ss_forecast
}
\description{
Example Stock Synthesis forecast file. Simulated data loosely based on the Pacific Hake stock assessment.
}
\keyword{datasets}
