% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_starter-data.r
\docType{data}
\name{ss_starter}
\alias{ss_starter}
\title{Example Stock Synthesis starter file. Simulated data loosely based on the Pacific Hake stock assessment.}
\format{
Named list created using \link[r4ss]{SS_readstarter}.
}
\usage{
ss_starter
}
\description{
Example Stock Synthesis starter file. Simulated data loosely based on the Pacific Hake stock assessment.
}
\keyword{datasets}
