% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osg_clean.r
\name{osg_clean}
\alias{osg_clean}
\title{This function cleans directories of given files and sub-directories}
\usage{
osg_clean(
  session = NULL,
  unix_name,
  login_node,
  rsa_keyfile = NULL,
  rsa_passphrase = NULL,
  remote_dirs,
  clean_files = NULL,
  clean_subdirs = NULL,
  verbose = TRUE
)
}
\value{

}
\description{
This function cleans directories of given files and sub-directories
}
