% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osg_download_ss_dir.r
\name{osg_download_ss_dir}
\alias{osg_download_ss_dir}
\title{This function downloads a directory of model runs to a local directory.}
\usage{
osg_download_ss_dir(
  session = NULL,
  unix_name,
  login_node,
  rsa_keyfile = NULL,
  rsa_passphrase = NULL,
  remote_dir_stem,
  remote_dirs,
  download_dir_stem,
  files_to_download = c("End.tar.gz"),
  untar_local = TRUE,
  clean_remote = TRUE,
  delete_remote = FALSE,
  verbose = TRUE
)
}
\description{
Options exist for cleaning the local and remote directory after download
}
