% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osg_execute.r
\name{osg_execute}
\alias{osg_execute}
\title{This function submits a job array to the Open Science Grid}
\usage{
osg_execute(
  session = NULL,
  unix_name,
  login_node,
  rsa_keyfile = NULL,
  rsa_passphrase = NULL,
  remote_submit_path = "scripts/condor_submit/",
  condor_submit_name = "condor.sub"
)
}
\arguments{
\item{condor_submit_name}{}
}
\description{
It is a wrapper around the condor_submit function
}
