% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osg_monitor.r
\name{osg_monitor}
\alias{osg_monitor}
\title{This function monitors an Open Science Grid job submission}
\usage{
osg_monitor(
  session = NULL,
  unix_name,
  login_node,
  rsa_keyfile = NULL,
  rsa_passphrase = NULL
)
}
\description{
It is a wrapper around the condor_q function
}
