% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osg_r_create.r
\name{osg_r_create}
\alias{osg_r_create}
\title{This function writes an R script which can be executed within the wrapper.sh to set-up and run diagnostics using existing r4ss code.}
\usage{
osg_r_create(
  session = NULL,
  unix_name,
  login_node,
  rsa_keyfile = NULL,
  rsa_passphrase = NULL,
  local_shell_path = NULL,
  remote_shell_path = "scripts/r/",
  diagnostic_type = NULL,
  overwrite = TRUE,
  verbose = TRUE,
  retro_n_years = NULL,
  r0_maxdiff = NULL,
  r0_step = NULL
)
}
\arguments{
\item{session}{ssh connection created by \link[ssgrid]{osg_connect}.}

\item{unix_name}{Character string giving OSG unix login name.}

\item{login_node}{Character string giving OSG login node (e.g., login05.osgconnect.net).}

\item{rsa_keyfile}{Path to private key file. Must be in OpenSSH format (see details). Default is NULL. See \link[ssh]{ssh_connect} for more details.}

\item{rsa_passphrase}{Either a string or a callback function for password prompt. Default is NULL. See \link[ssh]{ssh_connect} for more details.}

\item{local_shell_path}{Path to directory where condor_submit script is written. Defaults to \link[base]{tempdir}.}

\item{remote_shell_path}{Path to directory on OSG login node where condor_submit script is written.}

\item{diagnostic_type}{A character string specifying which diagnostic to run.
\describe{
\item{\emph{"01_run_retro"}}{Conduct a retrospective analysis of length \emph{retro_n_years}.}
\item{\emph{"02_run_R0profile"}}{Conduct a likelihood profile on R0 where the profile extends +/- \emph{r0_maxdiff} from the log(R0) maximum likelihood estimate at an interval specified by \emph{r0_step}.}
\item{\emph{"03_run_aspm"}}{Conduct age structured production (ASPM) and deterministic recruitment runs of the model. \emph{Warning: for the ASPM there are some hard coded settings that may not be appropriate for your model.}}
}}

\item{overwrite}{If the file given by \emph{file_name} exists in \emph{local_shell_path} or \emph{remote_shell_path} overwrite if TRUE.}

\item{verbose}{Boolean denoting if function details should be printed.}

\item{retro_n_years}{The number of years for the retrospective analysis.}

\item{r0_maxdiff}{The maximum distance from the log(R0) maximum likelihood estimate that the profile will extend.}

\item{r0_step}{The step size used in profiling (on a log scale).}
}
\value{
Returns 0 on exit.
}
\description{
This function saves a copy locally, uploads a copy to Open Science Grid (OSG), and modifies permissions so that it can be executed remotely
}
