% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osg_upload_ss_dir.r
\name{osg_upload_ss_dir}
\alias{osg_upload_ss_dir}
\title{This function uploads a directory of Stock Synthesis model runs to the OSG.}
\usage{
osg_upload_ss_dir(
  session = NULL,
  unix_name,
  login_node,
  rsa_keyfile = NULL,
  rsa_passphrase = NULL,
  local_dir_path,
  local_dir_names = NULL,
  remote_dir_path,
  files_to_upload,
  target_dir_path,
  target_dir_txt_name,
  verbose = TRUE
)
}
\arguments{
\item{rsa_passphrase}{}
}
\description{
Please see ?help for osg_connect and ssh::ssh_connect for more information.
}
