% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osg_wrapper_create.r
\name{osg_wrapper_create}
\alias{osg_wrapper_create}
\title{This function writes a bash shell wrapper script to be executed within each specified directory}
\usage{
osg_wrapper_create(
  session = NULL,
  unix_name,
  login_node,
  rsa_keyfile = NULL,
  rsa_passphrase = NULL,
  local_shell_path = NULL,
  remote_shell_path = "scripts/bash/",
  file_name = "wrapper.sh",
  wrapper_actions = NULL,
  overwrite = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{rsa_passphrase}{}
}
\description{
This function saves a copy of locally, uploads a copy to osg, and modifies permissions so that it can be executed remotely
}
